Some cleanup: added comments, moved helper functions, added get_proof_normal_index
