+++ src/mmr/mod.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod naive_merkle_mountain_ranges;
 pub mod naive_mmr_plonky2_verifier;
 pub mod naive_mmr_plonky2_verifier_1_recursion;
